Changed oxygen to use int 0-10000 avoid floating-point arithmetic. Also added a makePercent function to TextHelper to turn ints into human-readable percent strings (5500 -> 55.00%, 4839 -> 48.39%)

+++ include/Inventory.h
@@ -44,13 +44,13 @@ class Inventory : public ObjectContainer {
   * Returns the oxygen level of the spacesuit's internal oxygen tank, 0.00-1.00
   * @return a double showing the current oxygen level of the spacesuit.
   */
-  double getOxygen();
+  int getOxygen();
 
   /**
   * Sets the spacesuit's internal oxygen tank to the specified value.
   * @param oxygen the value to set the tank's oxygen level to.
   */
-  void setOxygen(double o);
+  void setOxygen(int o);
 
   /**
   * Adds the given object to the inventory, if object->isHoldable() is true.
@@ -60,7 +60,7 @@ class Inventory : public ObjectContainer {
 
  private:
   bool spaceSuitAcquired;
-  double oxygen;
+  int oxygen;
 };
 
 #endif // INVENTORY_H
